backend/requirements.txt
fastapi
uvicorn
pymongo
python-jose[cryptography]
pydantic
email-validator
python-dotenv
passlib
